VST3: move `GetFactoryPresetsPath` to `VST3Utils`
